refactoring to reduce duplication
