+++ jobs/estimate_job_counts.py
@@ -109,7 +109,7 @@ def main(
         r2=care_home_metrics_info["r2"],
         data_percentage=care_home_metrics_info["data_percentage"],
         model_version=care_home_model_info[-1],
-        model_name='care_home_with_nursing_historical_jobs_prediction',
+        model_name="care_home_with_nursing_historical_jobs_prediction",
         latest_snapshot=latest_snapshot,
         job_run_id=job_run_id,
         job_name=job_name,
@@ -131,7 +131,7 @@ def main(
         r2=non_residential_with_pir_metrics_info["r2"],
         data_percentage=non_residential_with_pir_metrics_info["data_percentage"],
         model_version=non_res_model_info[-1],
-        model_name='non_residential_with_pir',
+        model_name="non_residential_with_pir",
         latest_snapshot=latest_snapshot,
         job_run_id=job_run_id,
         job_name=job_name,
