+++ tests/unit/test_locations_non_res_feature_engineering.py
@@ -7,17 +7,11 @@ from pyspark.ml.linalg import SparseVector
 from pyspark.sql import SparkSession
 from pyspark.sql import functions as F
 from jobs import locations_non_res_feature_engineering
-from jobs.locations_non_res_feature_engineering import (
-    explode_column_from_distinct_values,
-    format_strings,
-    filter_locations_df_for_independent_non_res_care_home_data,
-    add_service_count_to_data,
-    add_date_diff_into_df,
-    add_rui_data_data_frame,
-)
+from jobs.locations_non_res_feature_engineering import filter_locations_df_for_independent_non_res_care_home_data
 from tests.test_file_generator import (
     generate_prepared_locations_file_parquet,
 )
+from utils.features.helper import add_date_diff_into_df
 
 
 class LocationsFeatureEngineeringTests(unittest.TestCase):
@@ -95,52 +89,6 @@ class LocationsFeatureEngineeringTests(unittest.TestCase):
 
         self.assertTrue(result.count() == 7)
 
-    def test_add_rui_data_data_frame(self):
-        cols = ["location", "rural_urban_indicator"]
-        rows = [
-            (
-                "1-10894414510",
-                {
-                    "year_2011": "(England/Wales) Rural hamlet and isolated dwellings in a sparse setting"
-                },
-            )
-        ]
-        df = self.spark.createDataFrame(rows, cols)
-        result = add_rui_data_data_frame(
-            df=df,
-            new_rui_col_name="rui_2011",
-            col_to_check="rural_urban_indicator.year_2011",
-            lookup_dict={
-                "indicator_1": "(England/Wales) Rural hamlet and isolated dwellings in a sparse setting"
-            },
-        )
-        result_rows = result.collect()
-        self.assertEqual(result_rows[0].indicator_1, 1)
-
-    def test_add_service_count_to_data(self):
-        cols = ["location", "services"]
-        new_col_name = "service_count"
-        rows = [("1", ["service_1", "service_2", "service_3"]), ("2", ["service_1"])]
-        df = self.spark.createDataFrame(rows, cols)
-
-        result = add_service_count_to_data(
-            df=df, new_col_name=new_col_name, col_to_check="services"
-        )
-        rows = result.collect()
-
-        self.assertTrue(
-            rows[0].asDict()
-            == {
-                "location": "1",
-                "services": ["service_1", "service_2", "service_3"],
-                "service_count": 3,
-            }
-        )
-        self.assertTrue(
-            rows[1].asDict()
-            == {"location": "2", "services": ["service_1"], "service_count": 1}
-        )
-
     def test_filter_locations_df_for_non_res_care_home_data(self):
         cols = ["carehome", "cqc_sector"]
         rows = [
@@ -160,86 +108,4 @@ class LocationsFeatureEngineeringTests(unittest.TestCase):
         self.assertTrue(result.count() == 1)
         self.assertEqual(
             result_row[0].asDict(), {"carehome": "N", "cqc_sector": "Independent"}
-        )
-
-    def test_format_strings(self):
-        string_1 = "lots of spaces"
-        string_2 = "CAPITALS"
-        string_3 = "Punctuation?!()"
-
-        self.assertEqual("lots_of_spaces", format_strings(string_1))
-        self.assertEqual("capitals", format_strings(string_2))
-        self.assertEqual("punctuation", format_strings(string_3))
-
-    def test_explode_column_from_distinct_values_returns_df_with_new_cols_and_prefix(
-        self,
-    ):
-        val_1 = "Glasgow"
-        val_2 = "London"
-        val_3 = "Leeds"
-
-        col_list = {val_1, val_2, val_3}
-
-        cols = ["locationid", "region"]
-        rows = [
-            ("1", val_1),
-            ("2", val_1),
-            ("3", val_2),
-            ("4", val_3),
-            ("5", val_3),
-        ]
-        df = self.spark.createDataFrame(rows, cols)
-
-        result = explode_column_from_distinct_values(
-            df=df, column_name="region", col_prefix="ons_", col_list_set=col_list
-        )
-        df_rows = result[0].collect()
-        list_created_of_new_cols = result[1]
-        list_created_of_new_cols.sort()
-
-        expected_returned_rows = [
-            {
-                "locationid": "1",
-                "region": "Glasgow",
-                "ons_leeds": 0,
-                "ons_london": 0,
-                "ons_glasgow": 1,
-            },
-            {
-                "locationid": "2",
-                "region": "Glasgow",
-                "ons_leeds": 0,
-                "ons_london": 0,
-                "ons_glasgow": 1,
-            },
-            {
-                "locationid": "3",
-                "region": "London",
-                "ons_leeds": 0,
-                "ons_london": 1,
-                "ons_glasgow": 0,
-            },
-            {
-                "locationid": "4",
-                "region": "Leeds",
-                "ons_leeds": 1,
-                "ons_london": 0,
-                "ons_glasgow": 0,
-            },
-            {
-                "locationid": "5",
-                "region": "Leeds",
-                "ons_leeds": 1,
-                "ons_london": 0,
-                "ons_glasgow": 0,
-            },
-        ]
-
-        actual_returned_rows = [row.asDict() for row in df_rows]
-
-        self.assertTrue(len(result[0].columns) == 5)
-        self.assertEqual(len(df_rows), len(rows))
-        self.assertListEqual(actual_returned_rows, expected_returned_rows)
-        self.assertEqual(
-            list_created_of_new_cols, ["ons_glasgow", "ons_leeds", "ons_london"]
         )
